commiting before adding date query
