+++ src/main/resources/templates/fragments/student_cal.html
@@ -2,7 +2,7 @@
 <html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns="http://www.w3.org/1999/html">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <title>SUTD ProfChoper</title>
 </head>
 <body>
