+++ src/main/resources/templates/student.html
@@ -90,7 +90,7 @@
     <p class="notitext">
         <span class="notistatus">Cancelled</span>
         4/12 prof oka from 3:30 - 4:00
-        <a class="notibtnbook btn btn-success" href="#calendar"> Book Another</a>
+        <a class="notibtnbook btn btn-success" href="#calendar">Book Another</a>
     </p>
 
     <!--if got msg-->
@@ -107,9 +107,11 @@
         <ul>
             <li class="prev">&#10094;</li>
             <li class="next">&#10095;</li>
-            <li>
-                <span th:text="${calendar.getHeader()}">Test</span><br>
-                <span style="font-size:18px">Week 12</span>
+            <li class="week-cal-header-text">
+                <span id="week-cal-header-date"
+                      th:text="${calendar.getHeaderDate()}">Test Date - Date20XX</span><br>
+                <span style="font-size:18px" id="week-cal-header-week"
+                      th:text="${calendar.getHeaderWeek()}">Week Test</span>
             </li>
         </ul>
     </div>
